% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BacVirsorter.R
\name{BacVirsorter}
\alias{BacVirsorter}
\title{VirSorter Table Preparation for CIRCOS}
\usage{
BacVirsorter(virsorter_file, karyotype_file)
}
\arguments{
\item{virsorter_file:}{The file path to the VirSorter output file containing scaffold, position and scoring information.}

\item{karyotype_file:}{The file path to the karyotype file containing scaffold and chromosomal group data.}
}
\value{
A data.frame with scaffold information mapped to CIRCOS formatting, including RGB color assignments and position data.
}
\description{
The BacVirsorter function processes VirSorter and karyotype data files to create a formatted table for CIRCOS visualization.
}
\examples{
virsorter_file <- read_table("data/virsorter.tsv", header = TRUE)
karyotype_file <- read_table("data/karyotype.tsv", col_names = FALSE)
formatted_circos <- BacVirsorter(virsorter_file, karyotype_file)
head(formatted_circos)

}
\author{
Dr. Abraham Cruz, Dr. Julio A. Hernandez-Gonzalez
}
